requirements.txt
Pillow==10.2.0
black==24.2.0
pylint==3.0.3
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
redis==5.0.3
PyYAML==6.0.1
pytz==2024.1
pytest==8.0.0
httpx==0.27.0  # For FastAPI testing